Split and tidy the board-array panelizer
commands/board_array.rs (4.5k lines) becomes a module directory:
vcut.rs (score lines, callouts, stroke-font labels), tooling.rs
(edge-rail tooling holes and fiducials), xml.rs (IPC-2581 patching
and generated-element serialization), tests.rs, with options,
validation, spec building, and the execute paths in mod.rs.

Four copies of fmt_num consolidate into utils::format; the panelizer
copper-layer match uses the shared layers::is_copper; the auto-plan
ranking and auto-options plumbing are flattened.
